+++ cmd/utils.go
@@ -343,7 +343,7 @@ func parseAtimeMtime(attr map[string]string) (atime, mtime time.Time, err *probe
 				return atime, mtime, probe.NewError(e)
 			}
 		}
-		atime = time.Unix(int64(atim), int64(atimnsec))
+		atime = time.Unix(atim, atimnsec)
 	}
 
 	if val, ok := attr["mtime"]; ok {
@@ -359,7 +359,7 @@ func parseAtimeMtime(attr map[string]string) (atime, mtime time.Time, err *probe
 				return atime, mtime, probe.NewError(e)
 			}
 		}
-		mtime = time.Unix(int64(mtim), int64(mtimnsec))
+		mtime = time.Unix(mtim, mtimnsec)
 	}
 	return atime, mtime, nil
 }
